Resolve some exceptions

+++ src/main/java/com/hikolu/crudrestapi/rest/DriversController.java
@@ -1,6 +1,8 @@
 package com.hikolu.crudrestapi.rest;
 
 import com.hikolu.crudrestapi.entity.Driver;
+import com.hikolu.crudrestapi.exception.BadRequestException;
+import com.hikolu.crudrestapi.exception.DriverNotFoundException;
 import com.hikolu.crudrestapi.service.DriversService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -27,33 +29,55 @@ public class DriversController {
     }
 
     // add mapping to GET "/{driverId}" and return a single driver based on id
-    // TODO add a check for driverId, weather it is int/is valid int if not throw exception
     @GetMapping("/{driverId}")
     public Driver findById(@PathVariable int driverId) {
+
+        Driver driver = driversService.findById(driverId);
+
+        if (driver == null) {
+            throw new DriverNotFoundException("Driver id not found, id - " + driverId);
+        }
+
         return driversService.findById(driverId);
     }
 
     // add mapping to PUT and return the updated driver entity
-    // TODO check weather the request body contains an id and if not throw an exception
     @PutMapping
     public Driver updateDriver(@RequestBody Driver driver) {
 
-        Driver dbDriver = driversService.save(driver);
+        if (driver.getId() == 0) {
+            throw new BadRequestException("Given request body doesn't contain id");
+        }
 
-        return dbDriver;
+        Driver dbDriver = driversService.findById(driver.getId());
+
+        if (dbDriver == null) {
+            throw new DriverNotFoundException("Driver not found, id - " + driver.getId());
+        }
+
+        return driversService.save(driver);
     }
 
     // add mapping to POST and return the new driver entity
-    // TODO check if it contains an id and if yes throw an exception
     @PostMapping
     public Driver addNewDriver(@RequestBody Driver driver) {
+
+        if (driver.getId() != 0) {
+            throw new BadRequestException("Given request body contains id");
+        }
+
         return driversService.save(driver);
     }
 
     // add a mapping to DELETE and return a string to display that operation was done
-    // TODO add a check for driverId, weather it is int/is valid int if not throw exception
     @DeleteMapping("/{driverId}")
     public String deleteEmployee(@PathVariable int driverId) {
+
+        Driver driver = driversService.findById(driverId);
+
+        if (driver == null) {
+            throw new DriverNotFoundException("Driver not found, id - " + driverId);
+        }
 
         driversService.deleteById(driverId);
